pathway 0.2: auto convert conics to quadratics
Conics are not supported in the public Android API. This new version
of Pathway automatically converts conic curves to approximated
quadratics curves. The default tolerance for the approximation
should be precise enough for most use cases.

This new version also offers new APIs to request conics to not
be converted so the path can be inspected "as-is".